Saving Courses State on local storage; Refactored Course to own file
