+++ app_feup/lib/controller/local_storage/AppUserDataDatabase.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'package:app_feup/controller/local_storage/AppDatabase.dart';
+import 'package:app_feup/model/entities/Course.dart';
 import 'package:app_feup/model/entities/Profile.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:tuple/tuple.dart';
